+++ tests/commands/sns-follow-neuron-unfollow.sh
@@ -1,0 +1,2 @@
+NEURON_ID=83a7d2b12f654ff58335e5a2512ccae0d7839c744b1807a47c96f5b9f3969069
+"$QUILL" sns follow-neuron $NEURON_ID --type upgrade-sns-to-next-version --unfollow --canister-ids-file sns_canister_ids.json --pem-file - | "$QUILL" send --dry-run