+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added `--confirmation-text` to `quill sns pay`. (#195)
+- Fixed `quill ckbtc update-balance` allowing the anonymous principal. (#191)
+- Added `disburse`, `disburse-maturity`, `split-neuron`, and `follow-neuron` to `quill sns`. (#191)
 - Added option to print DFN address for Genesis investors. (#184)
 - Updated to new ICRC-1 account ID format. (#190)
 
 ## [0.4.1] - 2023-03-23
 
-- Added release binaries for linux-gnu in addition to linux-musl on amd64.
+- Added release binaries for linux-gnu in addition to linux-musl on amd64. (#180)
 - Fixed `quill generate` requiring authentication. (#181)
 - Require an additional `--already-transferred` flag for the single-message form of `quill neuron-stake`. (#173)
 - Added `--disburse-amount` and `--disburse-to` to `quill neuron-manage`. (#171)